requirements.txt
# LangGraph / LangChain
langgraph>=0.6.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.3.0
langchain-text-splitters>=0.3.0

# 문서 처리
pdfplumber>=0.11.0

# 벡터 DB
chromadb>=1.5.0

# 뉴스 수집
duckduckgo-search>=6.0.0
feedparser>=6.0.0
python-dotenv>=1.0.0

# 유틸
python-dateutil>=2.9.0
tqdm>=4.66.0
pydantic>=2.0.0
